Replace schema files with scripts to create the needed views
Database tables can now be created by the Postfix Rest Server
(https://github.com/lyubenblagoev/postfix-rest-server) and the schema is
no longer necessary (and needs to be supported separately in case the
postfix-rest-server application's model changes). Users can create the
schema using the REST server and then use the scripts in the 'views'
directory to create the necessary views.